refactor: Extract the LogWriter from the log manager.
